requirements.txt
requests
pyperclip
prompt_toolkit
impacket
tabulate
colorlog
tmuxp
python-dateutil
